* automatic mode config added * added MySQL support * improved error handling * added saving map data and addresses between runs for supporting   append vs overwrite mode * added new main function of pymapmaker.py, enabling scheduling and looping * improved process cleanup * improved time keeping

+++ running.py
@@ -1,8 +1,16 @@
 import os
+import time
+
 import helpers
+from errors import InstanceError
+
+temp_dir = helpers.temp_dir
+if not os.path.exists(temp_dir):
+    os.makedirs(temp_dir)
+pid_file = os.path.join(temp_dir, 'pymapmaker.pid')
 
 
-pid_file = os.path.join('tmp', 'pymapmaker.pid')
+
 
 
 def check_running():
@@ -13,22 +21,35 @@ def check_running():
 def start_process():
     pid = str(os.getpid())
     if not check_running():
-        # os.system(run script here)
-        helpers.write_to_log("Program Started at " + helpers.start_time)
+        start_time = time.localtime(float(helpers.start_time))
+        start_time_string = time.strftime('%Y-%m-%d %H:%M:%S', start_time)
         helpers.write_to_log("Starting process with pid: " + pid)
+        helpers.write_to_log("Process started at " + start_time_string)
         f = open(pid_file, 'w')
         f.write(pid)
         f.close()
     else:
         helpers.write_to_log("Tried to start process but process is already running with pid: " + pid)
+        raise InstanceError('Instance already running with pid: ' + pid)
+
+
+def set_run_time():
+    # updates the last.time file for next run
+    last_run_file = open(os.path.join(temp_dir, 'last.time'), "w")
+    last_run_file.write(helpers.log_time)
+    last_run_file.close()
 
 
 def end_process():
-    end_time = helpers.get_time()
-    delta = int(end_time) - int(helpers.start_time)
-    helpers.write_to_log("Process took " + str(delta) + " seconds to finish.")
-    helpers.set_run_time()
-    os.remove(pid_file)
-
-
-
+    if check_running():
+        start_time_epoch = float(helpers.start_time)
+        end_time_epoch = float(helpers.get_time())
+        end_time = time.localtime(end_time_epoch)
+        end_time_string = time.strftime('%Y-%m-%d %H:%M:%S', end_time)
+        delta = end_time_epoch - start_time_epoch
+        helpers.write_to_log("Ending process at " + end_time_string)
+        helpers.write_to_log("Process took " + str(delta) + " seconds to finish.")
+        set_run_time()
+        os.remove(pid_file)
+    else:
+        helpers.write_to_log("Tried to end process but it appears the process was never started.")
